+++ Examify/backend/src/main/java/com/cis4000/examify/controllers/QuestionsController.java
@@ -1,28 +1,60 @@
 package com.cis4000.examify.controllers;
 
 import java.util.Optional;
+
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import com.cis4000.examify.models.Course;
 import com.cis4000.examify.models.Question;
+import com.cis4000.examify.repositories.CourseRepository;
 import com.cis4000.examify.repositories.QuestionRepository;
+
+import jakarta.servlet.http.HttpSession;
 
 import java.util.List;
 
 @RestController
 @RequestMapping("/api/questions")
-public class QuestionsController {
+public class QuestionsController extends BaseController {
 
-    private final QuestionRepository questionRepository;
+    @Autowired
+    private QuestionRepository questionRepository;
 
-    public QuestionsController(QuestionRepository questionRepository) {
-        this.questionRepository = questionRepository;
+    @Autowired
+    private CourseRepository courseRepository;
+
+    private boolean questionBelongsToUser(Question question, long userId) {
+        Optional<Course> courseOptional = courseRepository.findById(question.getCourseId());
+        if (courseOptional.isEmpty()) {
+            return false;
+        }
+
+        Course course = courseOptional.get();
+        Long courseUserIdObj = course.getUserId();
+        if (courseUserIdObj == null) {
+            return false;
+        }
+        long courseUserId = courseUserIdObj;
+        return courseUserId == userId;
     }
 
     @PostMapping
-    public ResponseEntity<String> createQuestion(@RequestBody Question question) {
+    public ResponseEntity<String> createQuestion(@RequestBody Question question, HttpSession session) {
         try {
+            Long userId = getUserIdOfSession(session);
+            if (userId == null) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                        .body("Error creating assignment: no user associated with current session");
+            }
+
+            if (!questionBelongsToUser(question, userId)) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                        .body("User does not have access to current course");
+            }
+
             questionRepository.save(question);
             return ResponseEntity.ok("Question created successfully");
         } catch (Exception e) {
@@ -32,7 +64,14 @@ public class QuestionsController {
     }
 
     @PutMapping("{id}")
-    public ResponseEntity<String> editQuestion(@PathVariable Long id, @RequestBody QuestionRequest questionRequest) {
+    public ResponseEntity<String> editQuestion(@PathVariable Long id, @RequestBody QuestionRequest questionRequest,
+            HttpSession session) {
+        Long userId = getUserIdOfSession(session);
+        if (userId == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                    .body("Error creating assignment: no user associated with current session");
+        }
+
         Optional<Question> questionOptional = questionRepository.findById(id);
 
         if (questionOptional.isEmpty()) {
@@ -40,6 +79,10 @@ public class QuestionsController {
         }
 
         Question question = questionOptional.get();
+        if (!questionBelongsToUser(question, userId)) {
+            ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                    .body("User does not have access to current question");
+        }
 
         question.setTitle(questionRequest.getTitle());
         question.setText(questionRequest.getText());
@@ -62,11 +105,22 @@ public class QuestionsController {
     }
 
     @DeleteMapping("{id}")
-    public ResponseEntity<String> deleteQuestion(@PathVariable Long id) {
+    public ResponseEntity<String> deleteQuestion(@PathVariable Long id, HttpSession session) {
+        Long userId = getUserIdOfSession(session);
+        if (userId == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                    .body("Error creating assignment: no user associated with current session");
+        }
+
         Optional<Question> questionOptional = questionRepository.findById(id);
 
         if (questionOptional.isEmpty()) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Question not found");
+        }
+
+        if (!questionBelongsToUser(questionOptional.get(), userId)) {
+            ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                    .body("User does not have access to current question");
         }
 
         questionRepository.deleteById(id);
